requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.1.4
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
gunicorn==21.2.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
mypy==1.14.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.8.2
yarg==0.1.9
